GOCBC-727: Modified all managers to not return errors.
Motivation
----------
Managers are created as pointer objects as opposed to actually
performing work when they are created.  They should reflect
this in terms of not returning errors.

Changes
-------
Modified the behaviour of manager creation such that they only
rely on having access to the cluster at operation execution
time, as opposed to at manager creation time.  Removed the
error return value from the manager creation methods as well.

Change-Id: I0c87c81af68a3eed0245f03b6bbcf2f2c4394cdf